Incorporating chad suggestions: check for duplicates - found in customers change the data type of price in products to numbers to prevent floating points unreliability

+++ set_up_scripts/create_db_scheme_tables.sql
@@ -26,7 +26,7 @@ create table if not exists customers (
 create table if not exists products (
   ProductID int PRIMARY KEY,
   Name varchar,
-  Price DOUBLE  
+  Price number(20,5)  
 );
 
 create table if not exists sales (
